Lots of things, including some major stuff
1) Fix the size decode logic to be more, well, logical

2) Finish implementation of the Vector48 decoder

3) Fix several bugs in the READ_WORD/READ_DWORD/READ_DWORD_X macros

4) Change the name of the vector-register ParameterType to VECTOR_REGISTER from VECTOR_REG

5) Clean up the ParameterTypes enum

6) Add a utility/spare "DATA" parameter type that should only hold strings (for now)

7) Refactor, cleanup and fix the to-string code in the instruction base-class

8) Update the CMake stuff to add the new dependencies.

+++ disasm_scalar48.cpp
@@ -97,7 +97,7 @@ namespace disasm {
                     do {
                         vc4_parameter d(ParameterTypes::REGISTER, (uint32_t)((insn | 0xffe0) & 0x001f));
                         vc4_parameter s(ParameterTypes::REGISTER,
-                                        ((insn >> 5) | 0xffe0) & 0x001f);
+                                        (uint32_t)(((insn >> 5) | 0xffe0) & 0x001f));
                         vc4_parameter u(ParameterTypes::IMMEDIATE, (uint32_t)(insn_arg));
                         scalar48_insn *rv = new scalar48_insn("add", "r{d}, " \
                                                               "r{s}, {u}");
